K8S-lab12/backend/requirements.txt
Flask==2.2.3
#psycopg2==2.9.6  # PostgreSQL adapter for Python
psycopg2-binary